+++ components/TagLabel/TagLabel.tsx
@@ -4,7 +4,7 @@ interface TagProps {
   textColor: string
 }
 
-export default function Tag({ tag, bgColor, textColor }: TagProps) {
+export default function TagLabel({ tag, bgColor, textColor }: TagProps) {
   return (
     <li
       className="rounded-[0.4rem] px-[0.6rem] py-[0.4rem] text-[1.2rem]"
