+++ components/index.ts
@@ -22,8 +22,12 @@ import DarkHeader from './Headers/DarkHeader/DarkHeader'
 import MyDashBoardList from './MyDashBoardList/MyDashBoardList'
 import Footer from './Footer/Footer'
 import InviteList from './InviteList/InviteList'
+import DashBoardCard from './DashBoardCard/DashBoardCard'
+import TagLabel from './TagLabel/TagLabel'
 
 export {
+  DashBoardCard,
+  TagLabel,
   SideMenu,
   DashBoardList,
   DashBoardListItem,
